Fixed entity class.
As outlined in the comments, Entities does not need an equals or hashCode method, since the database is handling these parts - we only use the entities as a mean for reading/writing content in the database. Business Logic is made on the DTO's.
The IntelliJ inspection rules should be followed, since they are there for a reason.
Dropped all copying, copying data in entities are not allowed, as also outlined in mails and trac. JPA implementations are using the Java Instrumentation framework, which allows memory monitoring, this is used to simplify the synchronization with the database, and means that by closing a connection the flushing will cause an unnecessary performance overhead.
Queries should not try to concatenate things directly, this should be left
Some of the tests rely in the equals in the entity - those tests should be dropped. We're relying on the equals in the DTO only.
The Client module, should *not* use any references to the entitites. The entitities are *only* suppose to be used in the Persistence & Core modules.