Corrected ignore messages.
@Michal; We should not have tests that are ignored. The tests in question all perform a check against the triggers in the database. However, triggers that perform input validation is a crazy contraption - since they either have to be made proper, i.e. cover *all* fields, or not at all.

The input validation checks should all be made in the API module, and also thoroughly tested there (positive, negative & border). It should not be possible for a request to get to the persistence level with invalid data. Hence, to avoid duplication of logic, the persistence layer should not need to perform any such checks.

Therefore, move these tests cases to the API module, and extend on them to cover all bases.