Added IWSEntity.
We need a simple way to verify if the Object at hand is an Entity or not. Further, the most commonly used method for all Entities is the "getId", so this method has been added as a required method. It also means that *all* objects must use a Long as primary key.