Added initial Monitoring suggestion and implementation.
The entire concept is based on Java Annotations & Reflection, and fulfills som of the requirement set forth regarding data protection mechanisms.
What is missing, is that the default MonitoringLevel to be used is defined in the Groups, and can be customized.
Further, the entire implementation so far is rather experimental, although it (theoretically) should work, the testing is not completed.