Authenticate added to Controllers & Services
The "new" way of dealing with the user information (The Authentication Object), is more or less completed now. The check is made in the Common Controller, result added to the other Controllers and send on to the Services. Now we have the user & group entity at hand for our further work :-)