Permission Check added.
The checks in the system to see if a user is allowed to perform a certain action is already in place, this is dealt with in the Controllers. However, the mechanism to verify if a user is permitted to access a given Object was not in place. The mechanism is simple, we compare the Group from the AuthenticationObject with the Group from the Object to be access. If not permitted, then a PermissionException is thrown.