Started corrections to the permissiom model.
The Permission model, is currently in a terrible state - that leaves the IWS near impossible to use. The work with correcting it all, and bringing it into a usable state is begun, and will be completed over the next few days.
Please note, that *any* change to the DB schema, will be rolled back ASAP, since it will most likely cause more harm than good - especially since there is currently under heavy development.