Cleanup of pom.
Dropped shading, not needed for the Client, since this is just suppose to provide the mechanism that wil l facilitate building configuration free apps based on IWS.
Dropped unused dependencies. Since we're now using XML Config, the PostgreSQL variant is discarded. It is the long-term goal to use JacaConfig, but the problems with Transactions are still lurking. It was not until Spring 3.1 that it was enabled, and there's still some bugs in Spring regarding this.