Replaced spring configuration with the Client module.
The FitNesse module should use the Client for all IWS communication, so we have a single location, where all changes are controlled, that relates to IWS & Spring configurations.
However, the recent changes in the Client module (moving from JavaConfig to XMLConfig), means that the FitNesse module builds, the JUnit tests runs - but the FitNesse tests doesn't!