fix testProcessOfferCreateFullOffer, refs #110
Changes in Offer.equals had to be made as java.util.Date is a time object and in database only date part is persisted.
As a result, after fetching object back from the database all date fields differ because of the Time part.
If we don't want to rely on external libraries in the API a new Comparator class seems to be the best solution.
It doesn't implement Comparator<?> so probably the name should be changed to avoid confusion.