Class & package renaming (refs #134)
The old package name was "notofications", it was renamed to "notifications".
The primary Notification implementation has been extended to implement the interface. The "old" methods are deprecated, since a more general setup is required. If we don't do it generally, it means that we have to write notify methods for each case, rather than leaving to the Objects that wishes to be notified, and thus lower the complexity.