Replaced Date with JodaTime for Offers.
The problem with needing a Date only value (i.e. no timestamp), required a simple solution. The Java API falls short on a nice solution, and JodaTime has been mentioned for integration into Java8, so I've taken the liberty to add it now - and thus fix our problems.
A Couple of further ToDo's have been added to the Offer. As I noticed a couple of fields were either missing or dropped. An explanation would be nice, if no fixing is necessary.