Completed Session Handling.
The sessions are not finalized. Before any operations can begin, a createSession must be made. It is then possible to verify the current session (and thus retrieve session data), store session data and once the session is over - deprecate it.
Please note, that a block has been added, so it is only allowed to have a single active session at the time. This was decided, so two concurrent sessions cannot screw up each others session data.