Added initial Permission Data.
The structure for the first couple of Permission checks is in place. Groups and users have been added, and existing have been removed. The new user accounts are: Austria, Croatia, Denmark, Germany and Poland, with password set to the same, i.e. username Austria -> password Austria.
The Notification has been slightly altered, I've added a Notification Spy to the Client, so it is possible via our tests to read out the notifications delivered.