Made Authorirization type specific.
Due to problems with the Permission mode, the Authorization Object containing GroupType & Permission was using Strings - it has now been reverted back to the original Objects (Enums) again.
Consolidation of Permissions is also being undertaken, so actions that are different, but reflect the same permission, are joined into one.
Started elaborating on the missing JavaDoc, still miles to go :-(