Changed Refno verification logic.
During the Skype Chat on 2012-11-05, it was agreed that the check for the refno should be simplified in the DTO validation, and an additional check should be made before persisting it, once we have the actual user/group information, and thus know the actual country that it should have.
Strange error on the FitNesse page, required me to change the user that was used in that test.