Corrections to Student DTO (refs #179).
I've corrected the DTO class, so it follows the norm. Please don't look at the Offer, since this class is rather inconsistent. Instead, use some of the other DTO's.
I've removed the Test Utility. This is a rather stupid way to go, since it obscures your tests, that you cannot see how things are generated. I have never been a fan of having to read multiple classes to understand a test. Even if it means that you have duplication of code. When you write a test, all information should be immediately present.
The Test itself is simplified, please look at how I did it.
The DateTime field for creation is removed from the DTO, it is a standard DB field, however, we only expose information in the API module, that is needed or required.