Consolidating API for User management (refs #100, #111).
The big problem with managing users, is that we need some kind of functionality, that allows managers to control individual user accounts, but at the same time also allow users themselves to update their own data. This resulted in a conflict with how we are trying to generalize the permission check. So instead, the checks and validations are moved into the implementing service.
Note this commit, is just the API changes. Not the implementation. As the implementation is independent of the API.