Extended the tables with constraint names.
When working with databases, it is often annoying that constraints are anonymous, i.e. given a random value, so you need to dig into them to check what their purpose is. Hence, I've extended some of the tables with the required constraints. Note, indexes are not yet added, mostly because the default indexes (primary key) is always present, and as long we don't have all the queries in place - it is hard to tell what indexes is required or not.