Added Notifications Period to the User Entity (refs #134).
We still have not reached a consensus regarding when notifications should be delivered. So I've extended the database and User Entity with a string field to allow an arbitrary value to be interpreted - this means that we can allow the "old" values such as "immediately", "hourly", "daily" and also combine it with a different type of notifications like a Cron version "* * * * *" for immediately or "0 8,12,16 * * 1-5" to mean that messages should be send on weekdays (Monday to Friday) at 8:00, 12:00 and 16:00.