Using generic error message for requests made with Null Objects.
It has bothered me for ages, that I couldn't come up with a better error message. So to at least avoid that we see crappy messages. a new Generic message was added. and our verify method uses varargs, so it is possible to provide an alternative if we need it.