Added reset Session (refs #100).
To ensure that it was possible to re-connect to the system, we needed a mechanism to reset dead sessions. As the IWS only allows that a user has 1 active session. With this update, the functionality has been added. Note, that the reason for the many files that are touched - is that the API Access interface is used many places, and that we needed to add a Notification Message Type to the sending of notifications.
Additional, this is a prototype of the new functionality, it is not yet tested - but only added (without breaking existing tests), so it can as quickly as possible be added to the IW4, in preparation for the AC. Tests will be added later.