Extended Offer saving with ExternalId (refs #221).
With this change, the Offer logic also uses the ExternalId, like other parts of the system. Creating Offers is unaffected, but updating will only be possible of the ExternalId & RefNo both are present and match.
Some of the tests in the system needed updating, and a single test has been ignored for now - since it required a complete rewrite - the usage of Mocking, combined with the logic should be re-thought.
This update will only affect the bug from Trac task #221, it does not contain the logic to specify which Offer to read, from a FetchOffer request. The FetchOffer request should be redone, since there are aspects of it that have not been updated since we added the Permission Module.