Added Parallel execution of our JUnit tests.
By running the tests in parallel, we will ensure that the system is less prone to Thread related errors. Note that the parallel execution requires that the Spring Beans are being created on a per-thread basis - as the Spring overhead is vastly greater than the test execution time, it means that the tests is becoming slower! If this is a problem, please make a local change where the value is replaced. The information is documented in the Parent pom. However, please do not check in the change.