Updated dependency versions.
Using the latest (and hopefully greatest) version of the different dependencies. I've also made some updates to the parallel execution of the tests, and by default it is now disabled, since Spring & Embedded databases seems to have problems :-( On real databases, I have so far not experienced any problems, but for our development, we need the HSQLDB in-memory database.