Fixed EJB problems...
The Glassfish problem has been boiled down to a problem with legacy code from version 2! This link [1] will explain it in details, please look under "To Disable GlassFish Server v2 Vendor-Specific JNDI Names". I understand the problem, but find the cause totally stupid! When referring to an injection, but specifying the specific interface, this legacy code sprung into effect and caused the problems.
The Remote interfaces have been removed, I found a way to add it to the API module, which was my original intention, but never did due to a conflict in the Persistence module, this means that we no longer require the marker interfaces in the EJB module, so I've moved the Beans up one package, so they now appear as first-class citizens in the EJB module.
Note, that I have not tested the fix with a database - so although all builds are successful, and I can deploy & redeploy without any problems, I cannot guarantee that it works.

[1] http://docs.oracle.com/cd/E18930_01/html/821-2416/gglpq.html