Extended all Offer Queries with Group (#100, #192).
The system only checks the following automatically: Request validity, Authentication of the User and if the User is Authorized. The result from this check in the Controllers will lead to a Service request with a new Object called Authentication, which contains the active Entities for the User and the current Group. It is then left to the developer to use this information to provide the correct checks in the further logic. The reason for this, is that the logic will have many different parameters that needs to taken into account.
The Queries in the OfferEntity have been extended with the Group check.
The FetchType "Owned" have been removed, since the "All" serve the same purpose, meaning - returning the list of current Offers belonging to the Group. No Group should be allowed to view other countries Offers, unless they have been shared.
Note; Since the FetchType enum has been changed, this will most like cause that the FE will not compile, please accept my apologies and correct the FetchType to "All".