Added Username & alias to User Object (#100, #267).
The User Object primarily serves the same purpose as the Profile Object in IW3, meaning that it holds the essential system information for the user. All private information is stored in a Person Object, that is yet to be clarified.
The Username is the users e-mail address - and the alias is the generated e-mail address on the system itself. The Alias cannot be altered, neither can the users first and last names.