Added new assertions, which throws IllegalArgumentException (#100)
Standard best practice, fail early, faily hard. Meaning, that the Object verification should be kept as simple as possible, and rather the actual checks should be moved into the setters, so invoking the setters will result in IllegalArgumentException's, if the provided data is illegal. This way, we'll minimize the amount of illegal Objects that the IWS needs to handle, since simple checks will result in client-side errors.
The concept was not earlier enforced. But as part of upgrading the JUnit tests, they now are. If this causes client-side problems, then they must be dealt with accordingly. The checks have not changed, only the place where they are enforced.