Don't just add things to the API!
There is enough Objects in the API, and they all have some relevance - adding new things that are not used is a very bad idea. If the TransportationType is not used in the API, then it should be moved out.
It is important to understand, that the IWS & IW4 are two separate projects, which share a common goal. However, as the IWS contains the public API, then only things that are important for this belongs here.