Moving Student Functionality into separate Interface (refs #100).
The Exchange module is slowly getting very crowded, so to avoid that it gets overly complex, the student functionality is now moved into a separate module. For now, the Interface alone is updated, later the actual move will be made. It is being done in smaller steps to ease the transition.