Code Review.
Various corrections made to the code. One important note - please stop all usage of mocking, one of the tests is failing because the assertions is made against a mock, which should *never* be used for validation, since you cannot be sure if you are testing that you wrote the mock correctly or the test correctly, hence - no mocking!