Migrated HSQLDB to PostgreSQL (refs #350).
The code is not yet tested, had a little problem, but given the late hour, I do not wish to try to debug it now. The tables are all working, all data is added, and most views are also working. The only thing that is not yet working, is the Employer view, since there is errors in it. This will have to be replaced with a proper table anyway.