fixes #334, #337, #338
The state of the offer is now available on the offer detail page.
A new functionality to see the number of countries an offer was shared to was added,
There are two new default sharing groups : "all" and "none"
Moreover, there is the begining of the functionality for the downloading of all offers as a csv which is implemented
Finally, the documents tab in the creation of an application was greatly improved but is not yet finalized.