fixes #342, ref #100
The documents tab of an application is now visualy correct.
A working version of the csv for the offers is now implemented, now it need to be changed in order to be the same as the one on  the old intraweb. There also needs to be unit testing implemented for it.