Changes to internal Notifications.
The "old" parts in the notification was fine for our early development stages, but it is now getting problematic, since we cannot properly work with templates. So, the current Notifiable methods have been deprecated and a new one has been added. Which together with the Notifications Implementations will be able to properly handle both lookup, message generation and sending.