Corrections to Database layer to get PostgreSQL to work (refs #368).
To get the Entities to work with both HSQLDB & PostgreSQL, it was necessary to beat Hibernate with a baseball bat, as the Hibernate developers have shit for brains! Sequences must be explicitly defined for PostgreSQL (though not for most other common DB's). And a few tiny mistakes were found.

Note, that the Employer view and the Employer logic is under massive re-write, and although most tests are working (only one is ignored), please expect that other tests may fail as part of this. Once the rewrite is completed, the PostgreSQL parts should work better, and the database schema should be far better suited for our September release.