More corrections to the database layer (refs #368).
All remaining ORM issues (mapping of database types vs entity types) are resolved. Both the Client & the Persistence module has been tested against a PostgreSQL database, and with a couple of notes, they work as can be expected.
Note regarding the OfferEntityTest; This test is testing if the database is working, like a database is suppose to work. Hard to explain, but the test is working under HSQLDB, but failing under PostgreSQL. The main reason is due to the braindead Hibernate developers, and the lack of support for reading uncommitted data in a transaction. Although I have tried to fiddle with the Spring isolation level, it lead to other problems which means that the only way to produce the expected errors, was by killing the transactions off, and flush between commits. However, killing off the transactions means that we also loose the wonderful little trick that a JUnit test is cleaning up after itself, by running in a transaction and rolling back after worth.
Note regarding the Client; Running the tests in the Client against a PostgreSQL will work perfectly for one complete run, since all tests are written correctly (i.e. independently) - however, running the same tests again a second time will cause all the expected problems that seems to be the subject for the OfferEntityTest. This means, that we do have a working PostgreSQL database, we can use it for our tests - and it should also be working for our Production Container, Glassfish.