Extended User processing.
I've added a primitive state machine for the user status changes. Further, it is not also possible to update the username. However, there is currently no linking to the PersonEntity for a User, nor is there a request method for actually updating the username. These are both still pending, as are the tests.