Added password to a users own update of username.
Same deal as with updating a password, the current is needed. Otherwise, a person abusing an active session may be able to request a username change, and then once the new username is confirmed, invoke a forgot password request. Following those steps an account would be hijacked. By enforcing that a user must also provide the password, the only thing that the system stores that is suppose to be a personal secret, is being used as authentication.
Note, that if an administrator attempts to update a username for a user, then the password is not required. The reason for this, is that an administrator is having the permissions to control the account, and thus can already do everything.