Preliminary User2Group functionality added (refs #100, #113).
Note, that the logic is rather complex, due to the many different internal rules regarding Groups and interaction. Rather than spending more time trying to get it all completed - I'm just checking in a preliminary version that should (untested :-( ) support the needs of our current system. This means that I can then focus on other more pressing matters and return to this one, once other tasks for the September Release are completed.