Corrected Offer refno format (refs #414).
Updated the format for our reference numbers from the old format (CCCC-YYYY-XXXX[-XX]) to the new format CC-YYYY-XXXXXX[-X]. CC is Country Code, YYYY is the enforced year, and the XXXXXX is the running number with the "-X" as optional value. The regex is as follows: [A-Z]{2}-\d{4}-\d{6}(-[A-Z0-9])?