Added (almost) all countries as test data.
A few Cooperating Institutions are missing, since our Test Data Generator will have to be way more complex if we are to control them all - and for testing it is not worth it.
Quite a few integration tests have been updated, to avoid problems with hardcoded Id's in the test.
Script 07-user-test-data.sql have been dropped from the list of test data in the Client, since it conflicts with the complete testdata script.