Completed corrections to Specializations & FieldsOfStudy (refs #416).
The Excel sheet was converted into an upper case listing that is now used as the default specializations, the description added is the text that the SID provided. The Fields of Study contained a couple of spelling mistakes. All of which has been corrected.
Only field missing was Dairy Farming - which was then added to the listing.
The list should now hopefully be correct with correct spellings and descriptions.