Bye bye SAR...
SAR's or Self Administrated Regions was added to IW3, to get around the problem with China. But instead it added a different problem, how to work with them. So, to makes everything easier, SAR's are dropped, and all Countries with SAR's (okay, only China), is reverting to just having National Groups. Thus, making things simpler :-)