Deprecation added in the Country Object (refs #100).
The Listname & National Secretary names must be removed from the Country Object, since they do not belong there. The main reason is, that these fields are strongly linked together wit the National Group, of which there can be multiple for any Country, if this Country is listed as a Co-operating Institution. For this reason, the List & NS can be confusing, since it is unclear which Group they belong to!

Fields are left for now, but should be removed once the Front-End (IW4 & PDF) both have verified the correct usage, i.e. via the Groups.