Cleanup of the file handling (refs #42).
Following a longer e-mail thread, it was made clear that the way files are handled in IWS vs. IW3, is different. In IW3, files were handled like in a "traditional" file system, whereas in IWS files are considered attachments. The File Entity/DTO has been updated to reflect this. If a folder structure is desired, it should be implemented in a different way.