Cleanup of the Address Object.
The Address has for a long time been annoying me, since it contained an ExternalId, but is considered an Internal Object, meaning that the API doesn't expose an Id to it. Now, I've removed the ExternalId and the lookup for this Object is purely based on the internal Id. All Objects requiring an Address will work with the address as an independent object, regardless of similarities.