Added support for deleting files (refs #42).
Rudimentary CRUD operations for files is now in place. The View for the files is still missing, as is support for the readData flag in the fetchFile request, currently all file data is returned. With the view we can control it better. But with this commit, we can use files.