Migration of basic tables completed (refs #505).
With this commit, the basic tables, users, groups, relations, are all migrated. With the required sub-tables and information. The migration is so far done with a simple JUnit test. So what is missing, is to make a more flexible tool that can be configured and also to extend the migration to cover the Offers and whatever else we need.