Initial version of Offer Statistics ready (refs #100, #547).
There is a couple of questions that needs clarification regarding how to generate statistics, or more likely, how to migrate existing data (mail discussion 2014-01-05). Until this is resolved, this commit is just considered preliminary, and will not be finalized until it can be properly tested.
Note, that most likely, all that needs changing is the two Database views.