Enable transition Cancelled - Applied for student application
Changing status works however it's not completed as there is a problem when only one country nominates a student and then cancel his/her application, both countries see status Cancelled. Then, if one uses this new transition, status of the application is changed but it's not propagated to the offer/offer group tables. This means, the offer owner still sees status Cancelled.