Initial mail migration completed (refs #505, #572).
With this, we're slowly crawling closer to completeness for this little tool. I've begun changing the configuration to use maven settings. This means, that you  can add values you your local settings.xml file, which will then be used for setting up our 3 data sources.
Migration of Mail is added to the mail datasource, including migration of the ncs mailinglist. So far, the mailinglist migration is just straightforward, users added. In IW3, more rules existed regarding who were on what mailinglist. However, for simplicity and to verify testing, it should hopefully be enough.