Corrected Maven configuration (refs #505).
Additional tried to make the migration in minor blocks to improve performance. However, it seems that Java is running with 90% CPU load, so further attempts at improving performance will require adding a ThreadPool to be processed. However, to get that to work with Spring will requires quite a bit of work, and this Migration has been dragging on for way too long now.