Added "Recent" to migration (refs #505).
With this change, it is possible to only migrate offers from the current Exchange Year (2014). This means that a recent migration is taking less than 10 minutes, compared to a full migration that takes hours. Note - I've added the Maven property value: "migrate.recent.offers.only" which is a boolean value, meaning that if set to "true", then we have the quick run, otherwise the slow.