Added DB Version table to IWS (refs #503).
With this, it is possible to save the current version and have an automatic update tool, which will update the DB if the version is incorrect. The version of the DB is linked together with the IWS Version from the Builds. This way, it is possible to trace back which DB version and IWS versions have been installed in production.