Corrections to the GroupNames (refs #612).
The Groupnames are now follow some strict conventions. Meaning, that Groups have names that depend on their `GroupType`. An Example; Austria Member Group have the Groupname "members", the full name "austria.members", the Austrian Staff now have the name "staff", and the full name "austria.staff".

The list name (for mailinglists), is also corrected, so it now depends on the type, what name it has. Both National & Member groups have the name "country" or "cooperating institution". Others have the same name as the full name, with whitespace replaced with underscores (still on the todo).

Missing, is JUnit testing, and inclusion in the Migration.