Fixed encoding problems (refs #621).
Stupid and annoying fix. The DB was using Strings to read, so I assume that Hibernate insisted on converting all unknown characters to some garbage. By reading the data in as blob's (byte arrays), I could omit the Hibernate part and then internally convert the Latin1 based Strings to UTF-8,which is what we're using internally.

Now I just hope that there are no other encodings in the DB. If so, then they will be considered lost! We cannot support two unknown encoding set at the same time.