Wanted to add more data but ...
... The fullname for Groups is a check we have to guarantee that the mailinglists are unique. The reason for having it there, and not on the list itself, is that Staff and Members both have the Committee Name as list, one with public the other with private. However, the information is appended to the Fullname.
The problem is, that for the Private Groups, it would be good to fill in the gaps also. Here we have the problem, that there exists 33 duplicate accounts. Most of these are simply stupid users who have taken over the NS role and instead of assigning it to the existing account, they take over the old account, and update the name!
This prevents us from filling in the fullname for Private Groups. However, rather than loosing this information, I'm committing it here for info.