Improved performance for Shared Offers (refs #648, #719).
Speed is now a factor 5 times faster, a single view replaced a bunch of DB requests. By testing with the HeavyLoad Test, it was possible to read 300 Sharings for Germany in ca. 15 seconds with 12 users in parallel requesting the result 25 times.
System is prepared for Pagination. Just need to get the FE to also update user frontend to support it.