Completed Validation of Student Application (refs #551).
Although validation should be a simple matter. The problem was that it was referring to the full Offer Object, but didn't provide it. Hence validation would fail! After discussing the matter with Matej, we agreed to drop the Offer Object and focus on the information required. Hence, a massive amount of changes on behalf of a single value :-/