Migration improvements (refs #505).
The migration suffered under the problem, that all records were read into one block and then processed sequentially in smaller blocks. With the changes, the all primary tables are now being processed in blocks, i.e. each block consists of a reading from IW3 and a migration + storing in IWS.