Completed Contact functionality (refs #748).
The Contact functionality is complete, it is possible to retrieve both the "overview" things, i.e. list of all International & Members Groups, sorted Alphabetically  - and from there fetch the individual sets of information, by either specifying a User or Group. The User Privacy field is also invoked, so only personal information is only showed when a user have changed the settings to Public.