Altered logic for deletion of file s(refs #631).
Added the same functionality as for #693, i.e. with a Storage Type, so if the StorageType is provided, the Group where the Application comes from, can also delete it.
However, the test in IW4 is wrong, it creates an offer as Austria, published an application as Croatia and attempts to delete the attachment as Austria, this is not permitted.