Added Transaction Control to Job Processing.
The problem we have had with the Database being unable to create multiple new Prepared Statements, may be related to the missing Transactional Control added to the Notification Manager Class. It is another attempt at trying to resolve our problems with the timer performing a rollback and killing all changes. Thus eventually killing itself.