Added static aliases (refs #612).
The Board has requested that some countries have a static alias. It is needed for those older Cooperating institutions which have been given status of associate members. Part of that change is that all committees are merged into a singular one with the country name as committee name. To ensure that the old lists are still working, the static aliases have been added.