Added support for retrieving all countries (refs #763).
The retrieval of countries was limited by the internal Countries View that was linking in the National committees to retrieve the NS & List. This meant that it was not possible to read out records for other types of information, like a full list of nationalities.
To fix this problem, the request has been expanded to support more different types.