Corrections to the WebService Annotations (refs #818).
With this correction, we do have a rudimentary WS available. When being deployed. The service cannot be access before the administrators have waved their magic hands on the server configuration.
There's still a todo with the WSDL, since this is currently not stored anywhere, but os relying on the autogenerated one. This has the problem that it prevents backward compatibility. But as this is the initial WebService, it should pose few of any problems.