New Inspection Rules.
The rules are are tried to be constructed so no style based issues are present, if some are, then please report it to me. No rules that generates false positives should be present - with the exception of rules that in compliance with the CERT rules for Java code.
The rules are made for IntelliJ 13, and will not work under old versions. The rules replaces earlier rules.