Removed legacy code (refs #648).
The Defensive / Deep Copying have been ignored for a while. Although it is generally best practice to enforce it, it was also the cause of some major performance problems.
Within the IWS, there was also the transformation layer between DTO's & Entities, which handled a part of this to avoid that changes to the DTO's triggered the Entity Instrumentation and thus caused unwanted DB updates.
Generally speaking, it is also not needed, since the clients for IWS is either trusted sources (tests & IW4) or WebService based. trusted sources should not be a safety concern and WebServices will always be made with a serialization/deserialization, meaning that all Object references are lost and thus it is not possible to externally change the state of Objects.