Added createStudent (refs #100, #773).
The new API method is present, and a simple test is added as a proof of concept. However, there seems to be a minor flaw in the code, which means that the verification of the existense of the account is failing. The exact reason is not found, and a bit puzzling, since the code is identical (actually calling) the same as for creating a user with the student flag set.