Bugfix, the lookup method for student groups was wrong.
The lookup is using JQL, but as JQL queries cannot be linked to the enumerated type. Hence we have to rely on correct typing. In the case of the Student Group lookup, the GroupType was referenced with "STUDENTS", but should correctly be "STUDENT" (i.e. no S in the end).