Minor corrections to mail migration.
Since the migration has to be re-run a few times until the new User Group interface is ready in the IW4, we can just as well try to weed out the few issues left. With this update, the NC's mailing list is having the prefix set properly, and the public lists for international groups is honoring the on_public_list flag.