Added Permission functionality (refs #787).
I have added some functionality to help with checking if a User is permitted to perform a given Action. I haven't had time to write tests for the methods, but they will hopefully give some ideas about how to do things correctly.
The former way of working with Permissions in IW4 is terrible - with this functionality, we can remove most of the crap from the system, and disregard the Role or Group specific checks and instead look at the Permissions, as intended.