Added fix for Sessions (refs #846).
The correction is simple - the deprecated field may not be null. However, a timestamp field to check against is always stupid since there can be problems for the precision or format. So I converted the field to a String and have set the default value.
The update script has also been tested and verified against a recent dump from the production database.