Corrections to Glassfish & JBoss configuration (refs #867).
With this correction, both Glassfish & JBoss (WildFly) works with the same Persistence.xml file. It had to be altered slightly, so it will work. I've added a sample domain.xml file for Glassfish to the accessories, and updated the Standalone XML file for JBoss.
Additionally, I've set the finalName for the Maven Build of IW4 to "intraweb", which will ensure that the WAR file is called intraweb, and thus will work under JBoss. For some reason, the context root is not read out and used when deployed :-( But this little trick, will ensure that the IntraWeb is always deployed with the correct context root.