Install script to create the databases with.
The new install script can be used to create both the IWS & Mail databases. The script also creates the roles required to work with the databases, and sets the username/password to the same as is in the domain.xml file for Glassfish.
Script is far from perfect, but it should hopefully be a start.