Correction to memory setup.
Pavel caught a mistake on the domain.xml file. Which is a duplicate -Xmx line, one with 2GB, another with 512MB. On the production server it was raised to 8GB, but this seems to cause other problems. So it is an attempt at avoiding too big problems and still have enough memory allocated to work - and 2GB is a fair limit on a 64 bit server.