Corrections to the CSV utils.
There were a couple of annoying issues with the CSV files. First, they were using a different encoding than the rest of the system. So now, they're using the same. Secondly, there seems to be a problem with the CSV files returned via the Servlet, since the ServletOutputStream is somehow removing all newlines. I have added some debugging information to see where the error is, and the data we're generating is with newlines, the data returned via the Stream is without!