Renamed Startup Bean to State Bean (refs #720, #900).
The State if the system is set at startup time by this Bean, and at the same time the Bean is starting a Timer service which runs as a Cron job every night at 2. The purpose of the Cronjob is to update Accounts to slowly get rid of trash. However, only the stubs for this is added, the actual logic is still pending.